add update users profile functionality, create request and response mapping

+++ src/main/java/com/epam/shevchenko/dao/BaseDAO.java
@@ -16,7 +16,7 @@ public interface BaseDAO<T> {
 	
 	List<T> showAll() throws DAOException;
 	
-	T getById(int id) throws DAOException;
+	T getById(long id) throws DAOException;
 	
 
 }
